unibot/requirements.txt
discord.py==2.1.0
openai
python-dotenv
requests